refactor: move non-admin auth checks into policies

+++ app/controllers/search_controller.rb
@@ -1,10 +1,10 @@
 class SearchController < ApplicationController
-  before_action :require_logged_in
-
   MAX_RESULTS = 8
 
   # GET /search/users.json?q=...
   def users
+    authorize :search
+
     q = params[:q].to_s.strip.delete_prefix("@")
 
     scope = User.discoverable.where.not(display_name: [ nil, "" ])
@@ -22,6 +22,8 @@ class SearchController < ApplicationController
 
   # GET /search/projects.json?q=...
   def projects
+    authorize :search
+
     q = params[:q].to_s.strip.delete_prefix("$")
 
     scope = Project.not_deleted
@@ -42,10 +44,5 @@ class SearchController < ApplicationController
   def avatar_for(slack_id)
     return nil if slack_id.blank?
     "https://cachet.dunkirk.sh/users/#{slack_id}/r"
-  end
-
-  def require_logged_in
-    return if current_user
-    head :unauthorized
   end
 end
